separate providers and models
