+++ examples/marvel/lib/src/screens/home/home.dart
@@ -1,87 +1,11 @@
-// ignore: import_of_legacy_library_into_null_safe
-import 'package:dio/dio.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../marvel.dart';
-import '../widgets/loading_image.dart';
-import '../widgets/marvel_logo.dart';
-import '../widgets/search_bar.dart';
-
-part 'home.freezed.dart';
-
-const kCharactersPageLimit = 50;
-
-@freezed
-class CharacterPagination with _$CharacterPagination {
-  factory CharacterPagination({
-    required int page,
-    String? name,
-  }) = _CharacterPagination;
-}
-
-class AbortedException implements Exception {}
-
-final characterPages = FutureProvider.autoDispose
-    .family<MarvelListCharactersReponse, CharacterPagination>(
-  (ref, meta) async {
-    // Cancel the page request if the UI no longer needs it before the request
-    // is finished.
-    // This typically happen if the user scrolls very fast
-    final cancelToken = CancelToken();
-    ref.onDispose(cancelToken.cancel);
-
-    // Debouncing the request. By having this delay, it leaves the opportunity
-    // for consumers to subscribe to a different `meta` parameters. In which
-    // case, this request will be aborted.
-    await Future<void>.delayed(const Duration(milliseconds: 250));
-    if (cancelToken.isCancelled) throw AbortedException();
-
-    final repository = ref.watch(repositoryProvider);
-    final charactersResponse = await repository.fetchCharacters(
-      offset: meta.page * kCharactersPageLimit,
-      limit: kCharactersPageLimit,
-      nameStartsWith: meta.name,
-      cancelToken: cancelToken,
-    );
-    return charactersResponse;
-  },
-  // When a page is no-longer use, keep it in cache for up to 5 minutes.
-  // After this point, if the list of characters is requested again, a new fetch
-  // will be performed.
-  cacheTime: const Duration(minutes: 5),
-);
-
-final charactersCount =
-    Provider.autoDispose.family<AsyncValue<int>, String>((ref, name) {
-  final meta = CharacterPagination(page: 0, name: name);
-
-  return ref.watch(characterPages(meta)).whenData((value) => value.totalCount);
-});
-
-@freezed
-class CharacterOffset with _$CharacterOffset {
-  factory CharacterOffset({
-    required int offset,
-    @Default('') String name,
-  }) = _CharacterOffset;
-}
-
-final characterAtIndex = Provider.autoDispose
-    .family<AsyncValue<Character>, CharacterOffset>((ref, query) {
-  final offsetInPage = query.offset % kCharactersPageLimit;
-
-  final meta = CharacterPagination(
-    page: query.offset ~/ kCharactersPageLimit,
-    name: query.name,
-  );
-
-  return ref.watch(characterPages(meta)).whenData(
-        (value) => value.characters[offsetInPage],
-      );
-});
+import '../../widgets/loading_image.dart';
+import '../../widgets/marvel_logo.dart';
+import '../../widgets/search_bar.dart';
+import 'home_models.dart';
+import 'home_providers.dart';
 
 class Home extends HookConsumerWidget {
   const Home({Key? key}) : super(key: key);
@@ -129,15 +53,14 @@ class Home extends HookConsumerWidget {
                   SliverPadding(
                     padding: const EdgeInsets.only(top: 10, left: 3, right: 3),
                     sliver: SliverGrid(
-                      gridDelegate:
-                          const SliverGridDelegateWithFixedCrossAxisCount(
+                      gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
                         crossAxisCount: 2,
                         childAspectRatio: 0.8,
                       ),
                       delegate: SliverChildBuilderDelegate((c, index) {
                         return ProviderScope(
                           overrides: [
-                            _characterIndex.overrideWithValue(index),
+                            characterIndex.overrideWithValue(index),
                           ],
                           child: const CharacterItem(),
                         );
@@ -147,8 +70,7 @@ class Home extends HookConsumerWidget {
                 ],
               ),
               floatingActionButton: FloatingActionButton.extended(
-                onPressed: () =>
-                    Navigator.pushNamed(context, '/characters/1009368'),
+                onPressed: () => Navigator.pushNamed(context, '/characters/1009368'),
                 label: const Text('Deep link to Iron-man'),
                 icon: const Icon(Icons.link),
               ),
@@ -158,14 +80,12 @@ class Home extends HookConsumerWidget {
   }
 }
 
-final _characterIndex = Provider<int>((ref) => throw UnimplementedError());
-
 class CharacterItem extends HookConsumerWidget {
   const CharacterItem({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final index = ref.watch(_characterIndex);
+    final index = ref.watch(characterIndex);
 
     final character = ref.watch(
       characterAtIndex(CharacterOffset(offset: index)),
